refactor(tui): extract shared story/activity widgets into screens/widgets.rs
Story rows (icon + pills + route lane), activity lines, status icons and
panel truncation move out of execute_dashboard/execute so the upcoming
workbench explorer can reuse them without copy-paste. story_list_item
gains a title_budget param for width-aware truncation.
