+++ crates/baro-tui/src/screens/execute_dag.rs
@@ -7,7 +7,7 @@ use ratatui::{
 };
 
 use crate::app::{App, LevelRunState, ReplanMark};
-use crate::screens::execute::status_icon_color;
+use crate::screens::widgets::status_icon_color;
 use crate::theme;
 
 pub fn render_dag_full(f: &mut Frame, app: &App, area: Rect) {
