+++ crates/baro-tui/src/screens/execute_dashboard.rs
@@ -9,7 +9,8 @@ use ratatui::{
     Frame,
 };
 
-use crate::app::{ActivityEntry, App, ReplanMark, StoryStatus};
+use crate::app::App;
+use crate::screens::widgets::{activity_line, story_list_item, truncate_for_panel};
 use crate::theme;
 
 pub fn render_dashboard(f: &mut Frame, app: &mut App, area: Rect) {
@@ -36,7 +37,7 @@ fn render_story_list(f: &mut Frame, app: &mut App, area: Rect) {
 
     if app.dag_levels.is_empty() {
         for story in &app.stories {
-            items.push(story_list_item(story, &app.push_results));
+            items.push(story_list_item(story, &app.push_results, 32));
         }
     } else {
         // Inner width of the story list block (panel - 2 borders - 1 for
@@ -59,7 +60,7 @@ fn render_story_list(f: &mut Frame, app: &mut App, area: Rect) {
 
             for story_id in level {
                 if let Some(story) = app.stories.iter().find(|s| s.id == *story_id) {
-                    items.push(story_list_item(story, &app.push_results));
+                    items.push(story_list_item(story, &app.push_results, 32));
                 }
             }
 
@@ -105,121 +106,6 @@ fn render_story_list(f: &mut Frame, app: &mut App, area: Rect) {
             .style(Style::default().fg(theme::MUTED));
         f.render_stateful_widget(scrollbar, area, &mut scrollbar_state);
     }
-}
-
-fn story_list_item(
-    story: &crate::app::StoryState,
-    push_results: &[(String, bool, Option<String>)],
-) -> ListItem<'static> {
-    let (icon, style) = match &story.status {
-        StoryStatus::Complete => ("✓", Style::default().fg(theme::SUCCESS)),
-        StoryStatus::Running => ("▶", Style::default().fg(theme::WARNING)),
-        StoryStatus::Failed => ("✗", Style::default().fg(theme::ERROR)),
-        StoryStatus::Retrying(_) => ("↻", Style::default().fg(theme::WARNING)),
-        StoryStatus::Skipped => ("⊘", Style::default().fg(theme::MUTED)),
-        StoryStatus::Pending => ("○", Style::default().fg(theme::MUTED)),
-    };
-
-    let duration = story
-        .duration_secs
-        .map(|d| format!(" ({}:{:02})", d / 60, d % 60))
-        .unwrap_or_default();
-
-    let push_indicator = if story.status == StoryStatus::Complete {
-        if let Some((_, success, _)) = push_results.iter().find(|(id, _, _)| id == &story.id) {
-            if *success {
-                Some(Span::styled(" ↑", Style::default().fg(theme::SUCCESS)))
-            } else {
-                Some(Span::styled(" ↑!", Style::default().fg(theme::ERROR)))
-            }
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-
-    // Truncate title with an ellipsis if it would exceed a reasonable
-    // budget for the 35%-width Stories column. The fixed prefix
-    // ("   {icon} {id}: ") is ~7-8 chars, suffixes (duration + pills +
-    // push indicator) can add ~15 more, so 32 chars for the title keeps
-    // the line below ~55 chars total — fits a 35% column on terminals
-    // ≥ 160 cols and degrades gracefully (List truncates, scrollbar
-    // works) on narrower ones.
-    let title = truncate_for_panel(&story.title, 32);
-
-    // Removed-by-replan stories read as struck plan surgery, not failure.
-    let title_style = if matches!(story.replan, Some(ReplanMark::Removed(_))) {
-        Style::default().fg(theme::MUTED).add_modifier(Modifier::CROSSED_OUT)
-    } else {
-        style
-    };
-
-    let mut spans = vec![
-        Span::raw("   "),
-        Span::styled(format!("{} {}: ", icon, story.id), style),
-        Span::styled(title, title_style),
-        Span::styled(duration, style),
-    ];
-    if let Some(indicator) = push_indicator {
-        spans.push(indicator);
-    }
-    spans.extend(story_pills(story));
-    if let Some(route) = &story.route {
-        spans.push(Span::styled(
-            format!("  {}", route),
-            Style::default().fg(theme::MUTED),
-        ));
-    }
-
-    ListItem::new(Line::from(spans))
-}
-
-/// Compact signal pills after the title — only rendered when the signal
-/// actually fired, so quiet stories stay quiet.
-fn story_pills(story: &crate::app::StoryState) -> Vec<Span<'static>> {
-    let mut pills: Vec<Span> = Vec::new();
-    let mut pill = |text: String, color: ratatui::style::Color| {
-        pills.push(Span::styled(format!(" {}", text), Style::default().fg(color)));
-    };
-
-    let retries = match story.status {
-        StoryStatus::Retrying(n) => n.max(story.retry_count),
-        _ => story.retry_count,
-    };
-    if retries > 0 {
-        pill(format!("↻{}", retries), theme::WARNING);
-    }
-    match story.critic_pass {
-        Some(true) => pill("✓critic".into(), theme::SUCCESS),
-        Some(false) => pill("✗critic".into(), theme::ERROR),
-        None => {}
-    }
-    if let Some(action) = &story.intervened {
-        let color = if action.contains("abort") { theme::ERROR } else { theme::WARNING };
-        pill("⚠stall".into(), color);
-    }
-    match story.merge {
-        Some(true) => pill("✓merged".into(), theme::SUCCESS),
-        Some(false) => pill("✗merge".into(), theme::ERROR),
-        None => {}
-    }
-    if story.replan.is_some() {
-        pill("✂replan".into(), theme::REPLAN);
-    }
-    pills
-}
-
-/// Truncate a string to `max` *characters* (not bytes), appending an
-/// ellipsis if it was shortened. Operates on `chars()` so multi-byte
-/// codepoints don't get cut in half.
-fn truncate_for_panel(s: &str, max: usize) -> String {
-    if s.chars().count() <= max {
-        return s.to_string()
-    }
-    let mut out: String = s.chars().take(max.saturating_sub(1)).collect();
-    out.push('…');
-    out
 }
 
 fn render_logs(f: &mut Frame, app: &App, area: Rect) {
@@ -384,60 +270,4 @@ fn render_logs(f: &mut Frame, app: &App, area: Rect) {
             f.render_stateful_widget(scrollbar, log_chunks[1], &mut scrollbar_state);
         }
     }
-}
-
-/// One Activity entry → a color-coded line. Icon carries the type signal;
-/// file changes + test verdicts color the whole line (diff-like). The panel
-/// is already scoped to one story, so no per-line agent prefix. System
-/// entries (replan/intervention/recovery/merge) get a ▸ accent prefix so run
-/// machinery stands apart from agent output.
-fn activity_line(e: &ActivityEntry, width: usize) -> Line<'static> {
-    if e.system {
-        let text_color = match e.kind.as_str() {
-            "replan" => theme::REPLAN,
-            "warn" | "conflict" => theme::WARNING,
-            "error" => theme::ERROR,
-            "merge" => theme::SUCCESS,
-            "verdict" => {
-                if e.ok == Some(true) { theme::SUCCESS } else { theme::ERROR }
-            }
-            _ => theme::ACCENT,
-        };
-        let budget = width.saturating_sub(2).max(8);
-        return Line::from(vec![
-            Span::styled("▸ ", Style::default().fg(theme::ACCENT)),
-            Span::styled(truncate_for_panel(&e.text, budget), Style::default().fg(text_color)),
-        ]);
-    }
-    let (icon, icon_color, text_color) = match e.kind.as_str() {
-        "tool_call" => match e.tool.as_deref() {
-            Some("bash") => ("$ ", theme::ACCENT, theme::TEXT),
-            Some("read") => ("· ", theme::TEXT_DIM, theme::TEXT_DIM),
-            _ => ("» ", theme::ACCENT, theme::TEXT),
-        },
-        "file_change" => match e.op.as_deref() {
-            Some("modify") => ("~ ", theme::WARNING, theme::WARNING),
-            _ => ("+ ", theme::SUCCESS, theme::SUCCESS),
-        },
-        "agent_msg" => ("› ", theme::TEXT_DIM, theme::TEXT_DIM),
-        "tool_result" => ("  ", theme::MUTED, theme::MUTED),
-        "test" | "verdict" => {
-            if e.ok == Some(true) {
-                ("✓ ", theme::SUCCESS, theme::SUCCESS)
-            } else {
-                ("✗ ", theme::ERROR, theme::ERROR)
-            }
-        }
-        "decision" => ("◆ ", theme::ACCENT, theme::TEXT),
-        "conflict" | "warn" => ("! ", theme::WARNING, theme::WARNING),
-        "error" => ("✗ ", theme::ERROR, theme::ERROR),
-        _ => ("  ", theme::TEXT, theme::TEXT),
-    };
-    // Truncate to the panel width (minus the 2-char icon) so long lines get a
-    // clean ellipsis instead of ratatui clipping them mid-word at the edge.
-    let budget = width.saturating_sub(icon.chars().count()).max(8);
-    Line::from(vec![
-        Span::styled(icon, Style::default().fg(icon_color)),
-        Span::styled(truncate_for_panel(&e.text, budget), Style::default().fg(text_color)),
-    ])
 }
